+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- Updated GitHub actions.
 - Updated `composer.json` with the PHP platform to 8.3.31.
 - `PdfWriter`: Defines class members as private.